requirements.txt
ib_async
pandas
python-dotenv
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
configparser